Reorganize into package. Add more details to installation instructions.
