+++ genome_loader/__main__.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python
-
 from pathlib import Path
 import sys
 import argparse
